% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{cumhaz.var}
\alias{cumhaz.var}
\title{Variance of the cumulative-hazard function}
\usage{
cumhaz.var(survi)
}
\arguments{
\item{survi}{a \code{survfit} object.}
}
\value{
A numeric vector.
}
\description{
\code{cumhaz.var} estimates the cumulative-hazard function variance using Greenwood's
formula for a \code{survfit} object containing estimated survival curves.
}
\examples{

library(survival)
fit <- survfit(Surv(stop, event) ~ 1, data=bladder)
cumhaz.var(fit)

}
\keyword{internal}

